added a hidden input to pass titlename to make deletion work in custom lists

+++ app.js
@@ -5,11 +5,14 @@ mongoose.connect("mongodb://localhost:27017/toDoListDB");
 
 const itemSchema = { item: { type: String, required: true } };
 
+const customSchema = {
+  name: { type: String, required: true },
+  items: [itemSchema],
+};
+
+const Custom = mongoose.model("Custom", customSchema);
+
 const Item = mongoose.model("Item", itemSchema);
-
-const item1 = new Item({ item: "type something" });
-
-const sampleItems = [item1];
 
 // ! -------------express boilerplate----------------
 
@@ -17,6 +20,7 @@ const express = require("express");
 const app = express();
 const port = 3000;
 const date = require(`${__dirname}/date.js`);
+let dayName = date();
 
 app.set("view engine", "ejs");
 
@@ -27,50 +31,79 @@ app.use(express.urlencoded({ extended: true }));
 app.use(express.static("public"));
 
 app.get("/", (req, res) => {
-  let dayName = date();
-
   Item.find((err, items) => {
-    if (items.length === 0) {
-      Item.insertMany(sampleItems, (err) => {
-        if (err) {
-          console.log(err);
-        } else {
-          console.log("sample items successfully saved in DB");
-        }
-      });
-      res.redirect("/");
-    } else {
-      // * in the below method current day is the marker in ejs file in the views folder
-      res.render("list", { listHeading: dayName, newItem: items });
-    }
+    // * in the below method current day is the marker in ejs file in the views folder
+    res.render("list", { listHeading: `${dayName} List`, newItem: items });
   });
 });
 
 app.post("/", (req, res) => {
   let newToDo = req.body.listAdd;
+  let titleName = req.body.list;
   // console.log(req.body);
-    // !------------collection for general list item----------------
-    const addedItem = Item({ item: newToDo });
-    addedItem.save((err) => {
+  const addedItem = Item({ item: newToDo });
+  // console.log(titleName);
+  // console.log(dayName);
+  if (titleName == dayName) {
+    addedItem.save();
+    res.redirect("/");
+  } else {
+    Custom.findOne({ name: titleName }, (err, foundItems) => {
       if (err) {
         console.log(err);
       } else {
-        console.log("new item successfully saved");
+        foundItems.items.push(addedItem);
+        foundItems.save();
       }
     });
-    res.redirect("/");
+    res.redirect(`/${titleName}`);
+  }
 });
 
 app.post("/delete", (req, res) => {
   const checkedItem = req.body.checked;
-  Item.findByIdAndRemove({ _id: checkedItem }, (err) => {
-    if (err) {
-      console.log(err);
-    } else {
-      console.log("deleted");
+  const listName = req.body.listName;
+
+  if (listName === `${dayName} List`) {
+    Item.findByIdAndRemove({ _id: checkedItem }, (err) => {
+      if (err) {
+        console.log(err);
+      } else {
+        console.log("deleted");
+      }
+    });
+    res.redirect("/");
+  }
+});
+
+app.get("/:newList", (req, res) => {
+  const newListName = req.params.newList;
+
+  Custom.findOne({ name: newListName }, (err, newList) => {
+    if (!err) {
+      if (!newList) {
+        // console.log("no such name found");
+        const newListItem = Custom({
+          name: newListName,
+          items: [],
+        });
+        newListItem.save((err) => {
+          if (err) {
+            console.log(err);
+          } else {
+            console.log("new list successfully saved");
+          }
+        });
+        res.redirect(`/${newListName}`);
+      } else {
+        // console.log("sorry the list name already exists........");
+        res.render("list", {
+          listHeading: `${capitalizeFirstLetter(newListName)} List`,
+          newItem: newList.items,
+        });
+      }
     }
   });
-  res.redirect("/");
 });
 
 app.get("/about", (req, res) => {
@@ -80,3 +113,11 @@ app.get("/about", (req, res) => {
 app.listen(process.env.PORT || port, () => {
   console.log(`Server started listening on port ${port}`);
 });
+
+// * function to make first letter capital
+function capitalizeFirstLetter(string) {
+  return string.charAt(0).toUpperCase() + string.slice(1);
+}
+function unCapitalizeFirstLetter(string) {
+  return string.charAt(0).toLowerCase() + string.slice(1);
+}
